requirements.txt
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
cssselect2==0.8.0
distro==1.9.0
Flask==3.1.2
fonttools==4.61.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
openai==2.8.1
pillow==12.1.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
pyphen==0.17.2
python-dotenv==1.2.1
sniffio==1.3.1
tinycss2==1.5.1
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
weasyprint==67.0
webencodings==0.5.1
Werkzeug==3.1.4
zopfli==0.4.0
